requirements.txt
web3
ipython
eth-brownie
pandas
google-cloud-storage
google-cloud-bigquery
